requirements.txt
Package      Version
------------ -------
kafka-python 2.0.2
pip          20.2.3
setuptools   49.2.1
